+++ src/Pagination/PaginationEntry.php
@@ -1,0 +1,40 @@
+<?php
+
+/**
+ * @copyright   ©2025 Maatify.dev
+ * @Library    maatify/data-repository
+ * @Project     maatify:data-repository
+ * @author      Mohamed Abdulalim (megyptm) <mohamed@maatify.dev>
+ * @since       2025-11-27 11:10
+ * @see         https://www.maatify.dev Maatify.com
+ * @link        https://github.com/Maatify/data-repository view project on GitHub
+ * @note        Distributed in the hope that it will be useful - WITHOUT WARRANTY.
+ */
+
+declare(strict_types=1);
+
+namespace Maatify\DataRepository\Pagination;
+
+class PaginationEntry
+{
+    public function __construct(
+        private int $page = 1,
+        private int $perPage = 10
+    ) {
+    }
+
+    public function getPage(): int
+    {
+        return $this->page;
+    }
+
+    public function getPerPage(): int
+    {
+        return $this->perPage;
+    }
+
+    public function getOffset(): int
+    {
+        return ($this->page - 1) * $this->perPage;
+    }
+}